differentiate windows tests
